Scale maxflow edges weights to reflect different cell sizes
Depending on the size of a cell the edge weights are scaled.
Weights from cells to the source or sink nodes are scaled by the area
of the cell and weights from cell to cell are scaled by the length of
the border between them. The weight from the implicir surrounding cell
to the sink is set to the maximum value possible to reflect that the
surrounding area is infinite.